Add files via upload
A much improved version of the Drumless demuxer script.

1. Changes the backend stem splitting tool to use python audio separator. This allows us to use the most up to date models such as BS-Roformer-SW for much improved quality and speed.

2. Now allows multiple and hi-res input file types.

3. Ported it to Python with improved error handling. This lays the ground work for a web based demuxer imterface. 

4. Decreased the amount of external tools in favour of python libraries. Note - I would shift the replaygain to ffmpeg but ffmpeg isn't as good as rsgain at this point for replaygain work.


Here is the old summary of the script:

This is quite a complicated script. It removes the drum track from a folder of FLAC music files using HTDEMUCS. But it does other things too.

Features include:

Prompt the user to whether we're processing an entire album or a collection (compilation of songs)
Uses HTDEMUCS in the highest quality setting (Demucs4_FT) and attempts to use Apple Silicon GPU encoding if possible
Uses ffmpeg to copy the id3 tags from the source files to the destination and then add useful other info such as Drumless to comments and Demucs model used with kid3.
Add compilation and album artist as Various Artists id3 if usr chose compilation in (1).
Uses kid3 to copy the cover art/picture from the source files to the desintation.
Uses ffmpeg to set id3 tags for Album or Compilation etc so music libraries can process them properly.
Uses RSGAIN to calculate and set ReplayGain tags to the output files. Noting RSGAIN doesnt support all Replaygain tags properly so use kid3 to write those. Calcuate Album gain if the user chose album in (1)
= The reason I wrote this was that I have a large collection of purchased music files in FLAC formats. I'm also a drummer. I want to be able to play along to real backing music. I don't want to pay a hefty ongoing subscription to commercial services such as Moises, and the other tools are missing features that I want/need.